monitoring/exporter/requirements.txt
prometheus_client~=0.20
google-cloud-monitoring~=2.17
fastapi~=0.111
uvicorn~=0.29
google-cloud-monitoring~=2.17 